fix: Move SortTab and Notification data classes
