+++ feature/notification/src/main/java/com/hyunjung/feature/notification/NotificationViewModel.kt
@@ -1,0 +1,1 @@
+package com.hyunjung.feature.notification
