+++ feature/notification/src/main/java/com/hyunjung/feature/notification/model/NotificationTabType.kt
@@ -1,0 +1,6 @@
+package com.hyunjung.feature.notification.model
+
+enum class NotificationTabType(val title: String) {
+    ALL("활동"),
+    READ("맞춤형")
+}
