+++ feature/notification/src/main/java/com/hyunjung/feature/notification/model/NotificationUiState.kt
@@ -1,8 +1,5 @@
 package com.hyunjung.feature.notification.model
 
-import com.hyunjung.feature.notification.viewmodel.NotificationItem
-import com.hyunjung.feature.notification.viewmodel.NotificationTabType
-
 data class NotificationUiState(
     val notifications: List<NotificationItem> = emptyList(),
     val selectedTab: NotificationTabType = NotificationTabType.ALL,
